lightningd: regroup hsm_secret password input logic
